# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Accounts","description":"Account management operations"},{"name":"Metrics","description":"Account metrics operations"},{"name":"Users","description":"User management operations"}],"paths":{"/users/{id}":{"get":{"tags":["Users"],"summary":"Get user by ID","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"put":{"tags":["Users"],"summary":"Update user","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["Users"],"summary":"Delete user","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}}}}},"/accounts/{id}":{"get":{"tags":["Accounts"],"summary":"Get account by ID","operationId":"getAccountById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}},"put":{"tags":["Accounts"],"summary":"Update account details","operationId":"updateAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccountRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}},"delete":{"tags":["Accounts"],"summary":"Delete account","operationId":"deleteAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}}}}},"/users":{"get":{"tags":["Users"],"summary":"Get all users","operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"post":{"tags":["Users"],"summary":"Create a new user","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/accounts":{"get":{"tags":["Accounts"],"summary":"Get all accounts","operationId":"getAllAccounts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"post":{"tags":["Accounts"],"summary":"Create a new account","operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/accounts/{id}/authorized-users":{"post":{"tags":["Accounts"],"summary":"Add an authorized user to the account","operationId":"addAuthorizedUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddAuthorizedUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/accounts/{id}/balance":{"patch":{"tags":["Accounts"],"summary":"Update account balance","operationId":"updateBalance","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBalanceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/users/{id}/balance":{"get":{"tags":["Users"],"summary":"Get user's total balance","operationId":"getUserBalance","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserBalanceResponse"}}}}}}},"/metrics/accounts":{"get":{"tags":["Metrics"],"summary":"Get account metrics by balance","operationId":"getAccountMetrics","parameters":[{"name":"greaterThan","in":"query","required":false,"schema":{"type":"number"}},{"name":"lessThan","in":"query","required":false,"schema":{"type":"number"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountMetricsResponse"}}}}}}},"/accounts/{id}/authorized-users/{userId}":{"delete":{"tags":["Accounts"],"summary":"Remove an authorized user from the account","operationId":"removeAuthorizedUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}}},"components":{"schemas":{"UpdateUserRequest":{"required":["email","name"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"}}},"AccountUserDTO":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"userName":{"type":"string"},"userEmail":{"type":"string"},"accountId":{"type":"integer","format":"int64"},"accountNumber":{"type":"string"},"balance":{"type":"number"},"role":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"accounts":{"type":"array","items":{"$ref":"#/components/schemas/AccountUserDTO"}}}},"UpdateAccountRequest":{"required":["accountNumber","balance"],"type":"object","properties":{"accountNumber":{"type":"string"},"balance":{"type":"number"}}},"AccountResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"accountNumber":{"type":"string"},"balance":{"type":"number"},"users":{"type":"array","items":{"$ref":"#/components/schemas/AccountUserDTO"}}}},"CreateUserRequest":{"required":["email","name"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"}}},"CreateAccountRequest":{"required":["accountNumber","balance","primaryUserId"],"type":"object","properties":{"accountNumber":{"type":"string"},"balance":{"type":"number"},"primaryUserId":{"type":"integer","format":"int64"}}},"AddAuthorizedUserRequest":{"required":["userId"],"type":"object","properties":{"userId":{"type":"integer","format":"int64"}}},"UpdateBalanceRequest":{"required":["balance"],"type":"object","properties":{"balance":{"type":"number"}}},"AccountSummary":{"type":"object","properties":{"accountId":{"type":"integer","format":"int64"},"accountNumber":{"type":"string"},"balance":{"type":"number"},"role":{"type":"string"}}},"UserBalanceResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"totalBalance":{"type":"number"},"accounts":{"type":"array","items":{"$ref":"#/components/schemas/AccountSummary"}}}},"AccountMetricsResponse":{"type":"object","properties":{"count":{"type":"integer","format":"int64"},"condition":{"type":"string"}}},"SuccessResponse":{"type":"object","properties":{"message":{"type":"string"},"timestamp":{"type":"integer","format":"int64"}}}}}}